requirements.txt
asgiref==3.2.10
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
cryptography==3.1
defusedxml==0.6.0
Django==3.1.1
django-allauth==0.42.0
django-cors-headers==3.5.0
django-rest-auth==0.9.5
djangorestframework==3.11.1
djangorestframework-jwt==1.11.0
gunicorn==20.0.4
h11==0.9.0
httpcore==0.10.2
httpx==0.14.3
idna==2.10
oauthlib==3.1.0
onesignal-sdk==2.0.0
Pillow==7.2.0
pycparser==2.20
PyJWT==1.7.1
PyMySQL==0.10.1
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
rfc3986==1.4.0
six==1.15.0
sniffio==1.1.0
sqlparse==0.3.1
urllib3==1.25.10
